+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### v0.0.2
+
+* Support for recording speech-to-text via Chrome SpeechRecognition.
+  * New events `recognise-interim` and `recognise`.
+  * Speech recognition is paused while the audio is played.
+
 ### v0.0.1
 
 * Initial release.